[Refactor] refactor application layer
