+++ application/src/main/kotlin/io/waterkite94/stalk/application/service/ContentLikeService.kt
@@ -1,6 +1,7 @@
-package io.waterkite94.stalk.application.usecase
+package io.waterkite94.stalk.application.service
 
 import io.waterkite94.stalk.application.port.LikePersistencePort
+import io.waterkite94.stalk.application.usecase.ContentLike
 import io.waterkite94.stalk.domain.model.vo.CommentLike
 import io.waterkite94.stalk.domain.model.vo.PostLike
 import org.springframework.stereotype.Service
