+++ application/src/main/kotlin/io/waterkite94/stalk/application/service/DeleteCommentService.kt
@@ -1,6 +1,7 @@
-package io.waterkite94.stalk.application.usecase
+package io.waterkite94.stalk.application.service
 
 import io.waterkite94.stalk.application.port.CommentPersistencePort
+import io.waterkite94.stalk.application.usecase.DeleteComment
 import org.springframework.stereotype.Service
 
 @Service
