+++ application/src/main/kotlin/io/waterkite94/stalk/application/service/DeletePostService.kt
@@ -1,6 +1,7 @@
-package io.waterkite94.stalk.application.usecase
+package io.waterkite94.stalk.application.service
 
 import io.waterkite94.stalk.application.port.PostPersistencePort
+import io.waterkite94.stalk.application.usecase.DeletePost
 import org.springframework.stereotype.Service
 
 @Service
