+++ application/src/main/kotlin/io/waterkite94/stalk/application/service/FindCommentService.kt
@@ -1,7 +1,8 @@
-package io.waterkite94.stalk.application.usecase
+package io.waterkite94.stalk.application.service
 
 import io.waterkite94.stalk.application.client.MemberServiceClient
 import io.waterkite94.stalk.application.port.CommentPersistencePort
+import io.waterkite94.stalk.application.usecase.FindComment
 import io.waterkite94.stalk.domain.model.vo.BoardComment
 import org.springframework.stereotype.Service
