+++ application/src/main/kotlin/io/waterkite94/stalk/application/service/FindPostService.kt
@@ -1,7 +1,8 @@
-package io.waterkite94.stalk.application.usecase
+package io.waterkite94.stalk.application.service
 
 import io.waterkite94.stalk.application.client.MemberServiceClient
 import io.waterkite94.stalk.application.port.PostPersistencePort
+import io.waterkite94.stalk.application.usecase.FindPost
 import io.waterkite94.stalk.domain.model.vo.BoardPost
 import org.springframework.stereotype.Service
