+++ application/src/test/kotlin/io/waterkite94/stalk/application/usecase/ContentLikeServiceTest.kt
@@ -1,6 +1,7 @@
 package io.waterkite94.stalk.application.usecase
 
 import io.waterkite94.stalk.application.IntegrationTestSupport
+import io.waterkite94.stalk.application.service.ContentLikeService
 import io.waterkite94.stalk.domain.model.vo.CommentLike
 import io.waterkite94.stalk.domain.model.vo.PostLike
 import org.junit.jupiter.api.DisplayName
