+++ application/src/test/kotlin/io/waterkite94/stalk/application/usecase/DeleteCommentServiceTest.kt
@@ -1,6 +1,7 @@
 package io.waterkite94.stalk.application.usecase
 
 import io.waterkite94.stalk.application.IntegrationTestSupport
+import io.waterkite94.stalk.application.service.DeleteCommentService
 import io.waterkite94.stalk.domain.model.vo.Comment
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatCode
